[refactor]: Separate profile components
